Introduce concept of lines with thier specific directions plus the CleanedPath and calculate accordingly.

+++ CleanerRobot/Entity/CleaningResult.cs
@@ -4,7 +4,7 @@ namespace CleanerRobot.Entity;
 
 public class CleaningResult
 {
-    public CleaningResult(DateTime requestedAt, int uniquePlacesCleaned, int commandCount, double duration)
+    public CleaningResult(DateTime requestedAt, long uniquePlacesCleaned, int commandCount, double duration)
     {
         RequestedAt = requestedAt;
         UniquePlacesCleaned = uniquePlacesCleaned;
@@ -14,7 +14,7 @@ public class CleaningResult
     
     public int Id { get; private set; }
     public DateTime RequestedAt { get; }
-    public int UniquePlacesCleaned { get; }
+    public long UniquePlacesCleaned { get; }
     public int CommandCount { get; }
     public double Duration { get; }
 }
